pyproject.toml
[tool.poetry]
name = "pairwisedistributiongathering"
version = "0.1.0"
description = ""
authors = ["d.ivakhnenko <d.ivakhnenko@ddos-guard.net>"]

[tool.poetry.dependencies]
python = "^3.8.5"
"infi.clickhouse-orm" = "^2.1.0"
tarantool = "^0.6.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
